JY.Qin task_01
#[Learning Experience](https://www.githubs.cn/)
**Personal feelings: <u>EXCITING & CHALLENGING</u>**
##Using a programme to briefly summarize it :
```python
print("To be honest, it\'s an amazing experience to learn Git and have a Github acountant of my own, alongside filled with hardships too.Nevertheless,I managed to get over all the problems by searching for information on the Internet and watching vedios at Bili Bili. For me,what matters is not to gain success,but to attain new approaches to thinking about questions.")
```
 *** 
###Expectations :
```c
 #include<stdio.h>
 int main()
 {
	int x;
	int n=0;
	printf("Please input an int at ramdom:\n");
	scanf("%d",&x);
	n++;
	x/=10;
	while(x>0){
	n++;
	x/=10;	
	}
	printf("Your int is composed of %d numbers!\n",n);
	return 0;
 }
```
 --- 

I want to have the capacity of counting bank notes just like the programme above in spite of there being a **limit** for it.   
Perhaps someday I will also gain the equivalent wealth,when that girl stands by me and parents smile on me.
 --- 
![](https://gss0.baidu.com/-vo3dSag_xI4khGko9WTAnF6hhy/zhidao/pic/item/cb8065380cd791235ae33b4cae345982b2b78083.jpg)